refactor(build): Use esbuild transpiler with rollup
The `typescript` rollup plugin does type checking, transpiling and emitting declaration files
together in a single process when used with rollup. This is particularly painful during the
dev mode as it takes above one minute to finish bundling for `core`.

This commit introduces esbuild plugin to handle transpiling while type checking and delaration
file emits are handled via a different process (in parallel) using typescript. The results are
encouraging and shared below for `core`.

New
`yarn build`: 33s
`yarn dev`: 12.5s for dist output (exluding type checking and declaration file emits)
`yarn dev` with changes: 4.5s for dist output (exluding type checking and declaration file emits)

Old
`yarn build`: 55s
`yarn dev`: 55s
`yarn dev` with changes: 20s

+++ app/scripts/modules/core/rollup.config.js
@@ -19,11 +19,13 @@ const fixTSPathRewrite = () => {
     name: 'fixTSPathRewrite',
     writeBundle: () => {
       const dts = path.join(__dirname, 'dist', 'index.d.ts');
-      const fixed = fs
-        .readFileSync(dts)
-        .toString()
-        .replace(/types="types"/, 'types="./types"');
-      fs.writeFileSync(dts, fixed);
+      if (fs.existsSync(dts)) {
+        const fixed = fs
+          .readFileSync(dts)
+          .toString()
+          .replace(/types="types"/, 'types="./types"');
+        fs.writeFileSync(dts, fixed);
+      }
     },
   };
 };
